+++ soridam-domain/src/main/java/sorisoop/soridam/domain/like/application/LikeQueryService.java
@@ -21,7 +21,7 @@ public class LikeQueryService {
 	private final LikeRepository likeRepository;
 
 	public LikeInfoDto getLikeInfo(User user, LikeType likeType, long targetId) {
-		int likeCount = likeRepository.countByLikeTypeAndTargetId(likeType, targetId);
+		long likeCount = likeRepository.countByLikeTypeAndTargetId(likeType, targetId);
 		boolean liked = likeRepository.existsByUserAndLikeTypeAndTargetId(user, likeType, targetId);
 
 		return LikeInfoDto.of(targetId, likeCount, liked);
